api/pyproject.toml
[project]
name = "wishlist-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.123.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "jsonify>=0.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy>=2.0.44",
    "sqlalchemy-utils>=0.42.1",
    "uvicorn>=0.38.0",
]
